posts.js 수정

+++ src/routes/posts.js
@@ -3,12 +3,38 @@ const psql = require("../../database/connect/postgre");
 // const mariadb = require("../../database/connect/mariadb");
 const modules = require("../module");
 
+//모든 게시글 읽기
+router.get('/', async (req, res) => {
+    const sql = "SELECT * FROM backend.post";
+    const result = {
+        "success": false,
+        "message": "",
+        "data": null
+    }
+
+    try {
+        const postData = await psql.query(sql);
+
+        if (postData.rows.length === 0) {
+            throw new Error('게시글이 존재하지 않습니다.');
+        }
+
+        result.success = true;
+        result.message = "정상적으로 데이터를 불러왔습니다.";
+        result.data = postData.rows;
+    } catch (err) {
+        result.message = err.message;
+    } finally {
+        res.send(result);
+    }
+})
+
 //게시글 작성
 router.post("/", async (req, res) => {
-    const { title, contents } = req.body;
+    const { title, contents, category } = req.body;
     const userIdx = req.session.idx;
-    const sql = "INSERT INTO post(user_idx, title, contents) VALUES(?, ?, ?)";
-    const params = [userIdx, title, contents];
+    const sql = "INSERT INTO backend.post(user_idx, title, contents, category) VALUES($1, $2, $3, $4)";
+    const params = [1, title, contents, category];
     const result = {
         "success": false,
         "message": ""
@@ -19,7 +45,7 @@ router.post("/", async (req, res) => {
             throw new Error("접근 권한이 없습니다.");
         }
 
-        await modules.query(sql, params);
+        await psql.query(sql, params);
 
         result.success = true;
         result.message = "게시글이 작성되었습니다.";
@@ -30,16 +56,15 @@ router.post("/", async (req, res) => {
     }
 })
 
-//게시글 읽기
+//특정 게시글 읽기
 router.get("/:idx", async (req, res) => {
     const postIdx = req.params.idx;
-    let sql = "SELECT * FROM post WHERE idx=?";
+    const sql = "SELECT * FROM backend.post WHERE idx=$1";
     const params = [postIdx];
     const result = {
         "success": false,
         "message": "",
-        "postData": null,
-        "commentData": null
+        "data": null
     };
 
     try {
@@ -47,26 +72,15 @@ router.get("/:idx", async (req, res) => {
             throw new Error("접근 권한이 없습니다.");
         }
 
-        const postData = await modules.query(sql, params);
+        const postData = await psql.query(sql, params);
 
-        if (postData.length === 0) {
+        if (postData.rows.length === 0) {
             throw new Error('해당 게시글이 존재하지 않습니다.');
         }
 
-        result.message = "게시글 읽기 성공";
-        result.postData = postData;
-
-        sql = "SELECT * FROM comment WHERE post_idx=?";
-
-        const commentData = await modules.query(sql, params);
-
-        if (commentData.length === 0) {
-            throw new Error('댓글이 존재하지 않습니다.');
-        }
-
         result.success = true;
-        result.message = "게시글, 댓글 읽기 성공";
-        result.commentData = commentData;
+        result.message = "정상적으로 데이터를 불러왔습니다.";
+        result.data = postData.rows;
     } catch (err) {
         result.message = err.message;
     } finally {
@@ -76,21 +90,26 @@ router.get("/:idx", async (req, res) => {
 
 //게시글 수정
 router.put("/:idx", async (req, res) => {
-    const { userIdx, title, contents } = req.body; //userIdx를 body로 받아오는 게 맞을까?
+    const { title, contents, category } = req.body;
+    const userIdx = req.session.idx;
     const postIdx = req.params.idx;
-    const sql = "UPDATE post SET title=?, contents=? WHERE idx=?";
-    const params = [title, contents, postIdx];
+    const sql = `
+        UPDATE backend.post
+        SET title=$1, contents=$2, category=$3
+        WHERE idx=$4 AND user_idx=$5
+    `;
+    const params = [title, contents, category, postIdx, userIdx];
     const result = {
         "success": false,
         "message": ""
     };
 
     try {
-        if (req.session.idx !== userIdx) {
+        if (!userIdx) {
             throw new Error("접근 권한이 없습니다.");
         }
 
-        await modules.query(sql, params);
+        await psql.query(sql, params);
 
         result.success = true;
         result.message = "게시글이 수정되었습니다.";
@@ -103,21 +122,21 @@ router.put("/:idx", async (req, res) => {
 
 //게시글 삭제
 router.delete("/:idx", async (req, res) => {
-    const { userIdx } = req.body;
+    const userIdx = req.session.idx;
     const postIdx = req.params.idx;
-    const sql = "Delete FROM post WHERE idx=?";
-    const params = [postIdx];
+    const sql = "Delete FROM backend.post WHERE idx=$1 AND user_idx=$2";
+    const params = [postIdx, 1];
     const result = {
         "success": false,
         "message": ""
     };
 
     try {
-        if (req.session.idx !== userIdx) {
+        if (!userIdx) {
             throw new Error("접근 권한이 없습니다.");
         }
 
-        await modules.query(sql, params);
+        await psql.query(sql, params);
 
         result.success = true;
         result.message = "게시글이 삭제되었습니다.";
